microservices_AG8/microservice2/requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
requests
